Update DataRepository/Collection nomenclature in glossary.
The old DataRepository concept maps mostly to the new Collection
concept, but in some LDM-556 requirements it seems to have been used
to mean "storage system" (i.e. Datastore backend) as well.
DataRepository is now used to mean an entity that has multiple
collections and potentially multiple storage systems, but generally
only one database, so this is perhaps the most important nomenclature
update.